% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connects to a running flowR server at the given host and port}
\usage{
connect(host = "localhost", port = 1042, blocking = TRUE)
}
\description{
Connects to a running flowR server at the given host and port
}
\seealso{
\code{\link[=send_flowr_request]{send_flowr_request()}}

\code{\link[=disconnect_flowr]{disconnect_flowr()}}
}
