% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{disconnect}
\alias{disconnect}
\title{Closes an active connection to a flowR server, if connected}
\usage{
disconnect(connection)
}
\arguments{
\item{connection}{The connection to close.}
}
\description{
Closes an active connection to a flowR server, if connected
}
\seealso{
\code{\link[=connect]{connect()}}

\code{\link[=send_request]{send_request()}}
}
