% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{disconnect_flowr}
\alias{disconnect_flowr}
\title{Closes an active connection to a flowR server, if connected}
\usage{
disconnect_flowr()
}
\description{
Closes an active connection to a flowR server, if connected
}
\seealso{
\code{\link[=connect_flowr_if_necessary]{connect_flowr_if_necessary()}}

\code{\link[=send_flowr_request]{send_flowr_request()}}
}
