% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr_local.R
\name{exec_docker_command}
\alias{exec_docker_command}
\title{Executes the given docker command in the given directory.
This function expects docker to exist on the system.}
\usage{
exec_docker_command(args, verbose = FALSE, cmd = "docker", background = FALSE)
}
\arguments{
\item{args}{The arguments to pass to the docker command, as a character vector.}

\item{verbose}{Whether to print out information about the commands being executed.}

\item{cmd}{The command to use for docker. Defaults to "docker".}

\item{background}{Whether the command should be executed as a background process.}
}
\value{
The return value of the call, which is the exit code if background is false, or the pid if background is true.
}
\description{
Executes the given docker command in the given directory.
This function expects docker to exist on the system.
}
