% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr_local.R
\name{exec_flowr_docker}
\alias{exec_flowr_docker}
\title{Executes a local version of the flowR CLI with the given arguments in the given directory.
This function expects docker to exist on the system.}
\usage{
exec_flowr_docker(
  docker_args,
  flowr_ver,
  flowr_args,
  verbose = FALSE,
  cmd = "docker",
  background = FALSE,
  std_out = TRUE,
  std_err = TRUE
)
}
\arguments{
\item{docker_args}{Additional arguments to pass to docker, as a character vector.}

\item{flowr_ver}{The version of flowR to use}

\item{flowr_args}{The arguments to pass to the flowR CLI, as a character vector.}

\item{verbose}{Whether to print out information about the commands being executed.}

\item{cmd}{The command to use for docker. Defaults to "docker".}

\item{background}{Whether the command should be executed as a background process.}

\item{std_out}{The std_out parameter passed to the sys call. See \code{\link[sys:exec]{sys::exec_background()}} and \code{\link[sys:exec]{sys::exec_wait()}} for more info.}

\item{std_err}{The std_err parameter passed to the sys call. See \code{\link[sys:exec]{sys::exec_background()}} and \code{\link[sys:exec]{sys::exec_wait()}} for more info.}
}
\value{
The return value of the \code{\link[=exec_node_command]{exec_node_command()}} call, which is the exit code if background is false, or the pid if background is true.
}
\description{
Executes a local version of the flowR CLI with the given arguments in the given directory.
This function expects docker to exist on the system.
}
