% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{exec_node_command}
\alias{exec_node_command}
\title{Executes the given Node subcommand in the given arguments in the given directory.
This function expects Node to have been installed using \code{\link[=install_node]{install_node()}}.}
\usage{
exec_node_command(
  app = c("node", "npm", "npx"),
  args,
  verbose = FALSE,
  base_dir = get_default_node_base_dir(),
  wait = TRUE
)
}
\arguments{
\item{app}{The node subcommand to run, which can be one of "node", "npm", or "npx".}

\item{args}{The arguments to pass to the Node command, as a character.}

\item{verbose}{Whether to print out information about the commands being executed.}

\item{base_dir}{The base directory that Node was installed in. By default, this uses the package's installation directory through \code{\link[=get_default_node_base_dir]{get_default_node_base_dir()}}.}

\item{wait}{Whether to wait for the command to finish before returning.}
}
\value{
The return value of the system2 call.
}
\description{
Executes the given Node subcommand in the given arguments in the given directory.
This function expects Node to have been installed using \code{\link[=install_node]{install_node()}}.
}
