% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr-package.R
\docType{package}
\name{flowr-package}
\alias{flowr}
\alias{flowr-package}
\title{flowr: Call flowR from R}
\description{
\code{flowr} is a thin, robust client for
\href{https://github.com/flowr-analysis/flowr}{flowR}, a static dataflow analyzer
and program slicer for R.  It talks to a flowR \emph{server} over a simple,
newline-delimited JSON protocol and can obtain that server through several
interchangeable engines (the JS+wasm bundle shipped in the package, a
self-contained downloaded binary, a private Node.js install, or Docker), so
that in the common case \code{slice()}, \code{query()} and friends "just work" without
the user installing anything by hand.
}
\details{
See \code{vignette("flowr")} to get started and \code{vignette("flowr-engines")} for
engine selection, the query API and the REPL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flowr-analysis/flowr-r-adapter}
  \item \url{https://github.com/flowr-analysis/flowr}
  \item \url{https://flowr-analysis.github.io/flowr/}
  \item Report bugs at \url{https://github.com/flowr-analysis/flowr-r-adapter/issues}
}

}
\author{
\strong{Maintainer}: Florian Sihler \email{florian.sihler@uni-ulm.de}

Authors:
\itemize{
  \item Florian Sihler \email{florian.sihler@uni-ulm.de}
}

Other contributors:
\itemize{
  \item flowr-analysis [copyright holder]
}

}
\keyword{internal}
