% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_json_string}
\alias{from_json_string}
\title{Create a R object from a JSON string. Can be used similar to \code{sprintf}.}
\usage{
from_json_string(x, ...)
}
\arguments{
\item{x}{The JSON string (with placeholders)}

\item{...}{The values to be inserted into the placeholders}
}
\value{
The R object
}
\description{
Create a R object from a JSON string. Can be used similar to \code{sprintf}.
}
