% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{get_default_node_base_dir}
\alias{get_default_node_base_dir}
\title{Returns the default node base directory to use when installing Node, which is the directory that the package with the given name is installed in.}
\usage{
get_default_node_base_dir(pkg_dir_name = "flowr")
}
\arguments{
\item{pkg_dir_name}{The name of the package to find the installation directory of. By default, this is "flowr", the name of this package.}
}
\value{
The default base directory to use when installing Node.
}
\description{
Returns the default node base directory to use when installing Node, which is the directory that the package with the given name is installed in.
}
