% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filetoken}
\alias{get_filetoken}
\title{Generates a token for a given file or content}
\usage{
get_filetoken(filepath = rlang::missing_arg(), content = rlang::missing_arg())
}
\arguments{
\item{filepath}{Path to a file}

\item{content}{Arbitrary content}
}
\value{
The token
}
\description{
Note that either \code{filepath} or \code{content} must be provided, but never both.
The token is then generated by hashing the file's content or the provided content,
depending on what was provided.
}
