% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_response}
\alias{handle_response}
\title{Checks if the given response was successful or not.}
\usage{
handle_response(res)
}
\arguments{
\item{res}{The response to handle}
}
\value{
A list with either the error's reason or the complete response
\describe{
\item{id}{The ID of the response (only in case of success).}
\item{res}{The complete response object (only in case of success).}
}
or
\describe{
\item{error}{The reason for the error (only present in case of an error).}
}
}
\description{
The function always returns a list with the response's ID. In case of an
error, the list also contains the errors reason. In case of a successful
response, the list also contains the complete response.
}
\seealso{
\code{\link[=send_request]{send_request()}}
}
