% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{install_flowr}
\alias{install_flowr}
\title{Installs the given version of flowR on the local system in the given directory.
This function expects Node to have been installed using \code{\link[=install_node]{install_node()}} prior.}
\usage{
install_flowr(
  flowr_ver,
  verbose = FALSE,
  base_dir = get_default_node_base_dir()
)
}
\arguments{
\item{flowr_ver}{The version of flowR to install.}

\item{verbose}{Whether to print out information about the commands being executed.}

\item{base_dir}{The base directory that Node was installed in, and where flowR should be installed. By default, this uses the package's installation directory through \code{\link[=get_default_node_base_dir]{get_default_node_base_dir()}}.}
}
\value{
The return value of the \code{\link[=exec_node_command]{exec_node_command()}} call.
}
\description{
Installs the given version of flowR on the local system in the given directory.
This function expects Node to have been installed using \code{\link[=install_node]{install_node()}} prior.
}
