% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{install_node}
\alias{install_node}
\title{Installs the given version of Node.js on the local system in the given directory.}
\usage{
install_node(node_ver, verbose = FALSE, base_dir = get_default_node_base_dir())
}
\arguments{
\item{node_ver}{The versino of Node to install}

\item{verbose}{Whether to print out information about the commands being executed.}

\item{base_dir}{The base directory to install Node in. By default, this uses the package's installation directory through \code{\link[=get_default_node_base_dir]{get_default_node_base_dir()}}.}
}
\description{
Installs the given version of Node.js on the local system in the given directory.
}
