% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{is_flowr_session}
\alias{is_flowr_session}
\title{Is this a flowR session?}
\usage{
is_flowr_session(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE}/\code{FALSE}.
}
\description{
Is this a flowR session?
}
\examples{
is_flowr_session(1)              # FALSE
\dontrun{
is_flowr_session(flowr_connect())
}
}
