% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{make_id_to_location_map}
\alias{make_id_to_location_map}
\title{Visits each node in the given AST and creates a map that maps node IDs to their locations in the code.}
\usage{
make_id_to_location_map(ast)
}
\arguments{
\item{ast}{The node or AST root to visit}
}
\value{
The ID-to-location map, where the keys are the node IDs and the values are the locations of the nodes.
}
\description{
Visits each node in the given AST and creates a map that maps node IDs to their locations in the code.
}
