% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{request_lineage}
\alias{request_lineage}
\title{Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-lineage-request}{"request-lineage"} request}
\usage{
request_lineage(con, filetoken, criterion, id = get_new_id())
}
\arguments{
\item{con}{The connection to the server}

\item{filetoken}{The filetoken retrieved by a previous "file-analysis"
request}

\item{criterion}{The criterion pointing to the object whose lineage should be
determined}

\item{id}{The id of the request}
}
\value{
A list containing the id and the response
}
\description{
Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-lineage-request}{"request-lineage"} request
}
\seealso{
\code{\link[=request_file_analysis]{request_file_analysis()}}

\code{\link[=connect]{connect()}}
}
