% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{request_query}
\alias{request_query}
\title{Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-query-request}{"request-query"} request}
\usage{
request_query(con, filetoken, query, id = get_new_id())
}
\arguments{
\item{con}{The connection to the server}

\item{filetoken}{The filetoken of the file to be sliced
(retrieved by a previous "file-analysis" request)}

\item{query}{The query to be executed}

\item{id}{The id of the request}
}
\value{
A list containing the id and the response
}
\description{
Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-query-request}{"request-query"} request
}
\seealso{
\code{\link[=request_file_analysis]{request_file_analysis()}}

\code{\link[=connect]{connect()}}
}
