% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repl.R
\name{request_repl}
\alias{request_repl}
\title{Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-repl-request}{"request-repl-execution"} request}
\usage{
request_repl(con, expression, id = get_new_id())
}
\arguments{
\item{con}{The connection to the server}

\item{expression}{The expression to be executed by the REPL}

\item{id}{The id of the request}
}
\value{
A list containing the id and the response
}
\description{
Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-repl-request}{"request-repl-execution"} request
}
\seealso{
\code{\link[=connect]{connect()}}
}
