% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{send_flowr_request}
\alias{send_flowr_request}
\title{Sends a JSON request to the running flowR server, if connected}
\usage{
send_flowr_request(command)
}
\description{
Sends a JSON request to the running flowR server, if connected
}
\seealso{
\code{\link[=connect_flowr_if_necessary]{connect_flowr_if_necessary()}}

\code{\link[=disconnect_flowr]{disconnect_flowr()}}
}
