% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{send_request}
\alias{send_request}
\title{Sends a JSON request to the running flowR server, if connected}
\usage{
send_request(connection, command)
}
\arguments{
\item{connection}{The connection to use when sending the request.}

\item{command}{The command to send, as a named list representing the key-value pairs to set in the JSON request.}
}
\description{
Sends a JSON request to the running flowR server, if connected
}
\seealso{
\code{\link[=connect]{connect()}}

\code{\link[=disconnect]{disconnect()}}
}
