% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{send_request_handle_response}
\alias{send_request_handle_response}
\title{Send the given request over the given connection and handle a possible error}
\usage{
send_request_handle_response(con, request)
}
\arguments{
\item{con}{The connection to use}

\item{request}{The request to send}
}
\value{
See \code{\link[=handle_response]{handle_response()}} for possible return values
}
\description{
Send the given request over the given connection and handle a possible error
}
