% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{visit_node}
\alias{visit_node}
\title{Visits the given node and all of their children, invoking the callback for each visited node}
\usage{
visit_node(node, callback)
}
\arguments{
\item{node}{The node to visit}

\item{callback}{The callback function to invoke for each node. The callback should return TRUE to continue visiting the children of the node, or FALSE to stop visiting the children of the node.}
}
\description{
Visits the given node and all of their children, invoking the callback for each visited node
}
