% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{visit_nodes}
\alias{visit_nodes}
\title{Visits the set of nodes and all of their children, invoking the callback for each visited node}
\usage{
visit_nodes(nodes, callback)
}
\arguments{
\item{nodes}{The list or array of nodes to visit}

\item{callback}{The callback function to invoke for each node. The callback should return TRUE to continue visiting the children of the node, or FALSE to stop visiting the children of the node.}
}
\description{
Visits the set of nodes and all of their children, invoking the callback for each visited node
}
