% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connects to a running flowR server at the given host and port}
\usage{
connect(host = "localhost", port = 1042, blocking = TRUE)
}
\arguments{
\item{host}{The host to connect to. By default, this is "localhost".}

\item{port}{The port to connect to. By default, this is 1042.}

\item{blocking}{Whether to use blocking mode for the connection.}
}
\description{
Connects to a running flowR server at the given host and port
}
\seealso{
\code{\link[=send_request]{send_request()}}

\code{\link[=disconnect]{disconnect()}}
}
