% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowr_local.R
\name{exec_docker_command}
\alias{exec_docker_command}
\title{Executes the given docker command in the given directory.
This function expects docker to exist on the system.}
\usage{
exec_docker_command(
  args,
  verbose = FALSE,
  cmd = "docker",
  background = FALSE,
  std_out = TRUE,
  std_err = TRUE
)
}
\arguments{
\item{args}{The arguments to pass to the docker command, as a character vector.}

\item{verbose}{Whether to print out information about the commands being executed.}

\item{cmd}{The command to use for docker. Defaults to "docker".}

\item{background}{Whether the command should be executed as a background process.}

\item{std_out}{The std_out parameter passed to the sys call. See \code{\link[sys:exec]{sys::exec_background()}} and \code{\link[sys:exec]{sys::exec_wait()}} for more info.}

\item{std_err}{The std_err parameter passed to the sys call. See \code{\link[sys:exec]{sys::exec_background()}} and \code{\link[sys:exec]{sys::exec_wait()}} for more info.}
}
\value{
The return value of the call, which is the exit code if background is false, or the pid if background is true.
}
\description{
Executes the given docker command in the given directory.
This function expects docker to exist on the system.
}
