% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_analysis.R
\name{request_file_analysis}
\alias{request_file_analysis}
\title{Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-analysis-request}{"file-analysis"} request}
\usage{
request_file_analysis(
  con,
  filepath = rlang::missing_arg(),
  content = rlang::missing_arg(),
  cfg = FALSE,
  id = get_new_id(),
  filetoken = get_filetoken(filepath, content)
)
}
\arguments{
\item{con}{The connection to the server}

\item{filepath}{A single or multiple file paths (must be visible to the server)}

\item{content}{The R code to be analyzed}

\item{cfg}{Weather to include the control flow graph in the response}

\item{id}{The id of the request}

\item{filetoken}{The filetoken of the file to be analyzed}
}
\value{
A list containing the id, filetoken and the response
}
\description{
Note that either \code{filepath} or \code{content} must be provided, but never both.
}
\seealso{
\code{\link[=connect]{connect()}}
}
