% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_analysis.R
\name{request_file_analysis}
\alias{request_file_analysis}
\title{Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-analysis-request}{"file-analysis"} request
Note that either \code{filepath} or \code{content} must be provided, but never both.}
\usage{
request_file_analysis(
  filepath = NULL,
  content = NULL,
  id = get_new_id(),
  filetoken = get_filetoken(),
  con = connect()
)
}
\arguments{
\item{filepath}{The path to the file (must be visible to the server)}

\item{id}{The id of the request}

\item{filetoken}{The filetoken of the file to be analyzed}

\item{con}{The connection to the server}
}
\value{
A list containing the id, filetoken and the response
}
\description{
Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-analysis-request}{"file-analysis"} request
Note that either \code{filepath} or \code{content} must be provided, but never both.
}
