% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{request_slice}
\alias{request_slice}
\title{Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-slice-request}{"request-slice"} request}
\usage{
request_slice(filetoken, criteria, id = get_new_id(), con)
}
\arguments{
\item{filetoken}{The filetoken of the file to be sliced
(retrieved by a previous "file-analysis" request)}

\item{criteria}{The slice's criteria}

\item{id}{The id of the request}

\item{con}{The connection to the server}
}
\value{
A list containing the id and the response
}
\description{
Sends a \href{https://github.com/flowr-analysis/flowr/wiki/Interface#the-slice-request}{"request-slice"} request
}
\seealso{
\code{\link[=initiate_file_analysis]{initiate_file_analysis()}}

\code{\link[=connect]{connect()}}
}
