% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{send_request}
\alias{send_request}
\title{Sends a JSON request to the running flowR server, if connected}
\usage{
send_request(connection, command)
}
\description{
Sends a JSON request to the running flowR server, if connected
}
\seealso{
\code{\link[=connect]{connect()}}

\code{\link[=disconnect]{disconnect()}}
}
