% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit.R
\name{visit_node}
\alias{visit_node}
\title{Visits the given node and all of their children, invoking the callback for each visited node}
\usage{
visit_node(node, callback)
}
\arguments{
\item{callback}{The callback function to invoke for each node}

\item{nodes}{The node to visit}
}
\description{
Visits the given node and all of their children, invoking the callback for each visited node
}
